Update recipe to use chromium m64 r517731.
This is a transient update of the current recipe to update chromium
to version m64 r517731 based on Igalia/ozone-wayland-dev.

The recipe will be updated further to allign better with OSSystem/master
branch and some of the changes are going to be merged soon.

major changes:
1) PACKAGECONFIG variables are different now. Mandatory variables for
AGL builds are use-gold and webrtc.
2) Build target is chromium-ozone-igalia now.