chromium: Update to 77.0.3865.75
This is a security fix release:
https://chromereleases.googleblog.com/2019/09/stable-channel-update-for-desktop.html

Changes:
* Disabled unbinding of libwebp. That is, some chromium components
depend on entire //skia that also depennds on libwebp. That might be ok,
but there are some targets, which are required to be built with host
toolchains. This results in pkg-config-native to be used and inability
to find libwebp packages. A bug to chromium will be submitted.

* Do not try to pack 22x22 logos as those are no longer used.
https://crrev.com/c/1671427

Test-built with:

* chromium-x11
  * thud, MACHINE="intel-corei7-64"
  * thud, MACHINE="intel-core2-32"
  * warrior, MACHINE="intel-corei7-64"
  * warrior, MACHINE="intel-corei7-64"
* chromium-ozone-wayland:
  * thud, MACHINE="qemuarm64"
  * thud, MACHINE="m3ulcb"
  * warrior, MACHINE="raspberrypi3"

Fixes #303.

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>
Signed-off-by: Maksim Sisov <msisov@igalia.com>